% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegie.R
\name{buildDB}
\alias{buildDB}
\title{Build database from legacy format.}
\usage{
buildDB(dsn, layer, dbfile, fidoffset = 0)
}
\arguments{
\item{dsn}{data source name}

\item{layer}{layer name}

\item{dbfile}{SQLite file name}
}
\description{
Build database from legacy format.
}

